pyproject.toml
[project]
name = "judgeval"
version = "0.0.5"
authors = [
    { name="Andrew Li", email="andrew@judgmentlabs.ai" },
    { name="Alex Shan", email="alex@judgmentlabs.ai" },
    { name="Joseph Camyre", email="joseph@judgmentlabs.ai" },
]
description = "Judgeval Package"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]
dependencies = [
    "langfuse==2.50.3",
    "litellm>=1.48.10",
    "python-dotenv==1.0.1",
    "fastapi>=0.115.6",
    "uvicorn>=0.34.0",
    "deepeval>=2.1.8",
    "supabase>=2.11.0",
    "requests>=2.32.3",
    "pandas>=2.2.3",
    "openai>=1.47.1",
    "together>=1.3.11",
    "anthropic>=0.43.1",
    "patronus>=0.0.17"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "pytest-mock>=3.14.0"
]

[project.urls]
Homepage = "https://github.com/JudgmentLabs/judgeval"
Issues = "https://github.com/JudgmentLabs/judgeval/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/judgeval"]
include = [
    "/src/judgeval",
    "/src/judgeval/**/*.py",
]

[tool.hatch.build]
directory = "dist"
artifacts = [
    "src/judgeval/**/*.py",
]